Move file store to shared package
Move FileStorage to shared package so that server may depend on
it without requiring bulky ML dependencies.
